checkpoint

+++ libx/stdlib.h
@@ -1,8 +1,9 @@
 #ifndef STDLIB_H
 #define STDLIB_H
+#include <stddef.h>
 
 axiom void *
-malloc(int size) [ .alloc result; ];
+malloc(size_t size) [ .alloc result; ];
 
 axiom void
 free(void *ptr) [ .dealloc ptr; ];
